frontend/requirements.txt
Flask>=2.2
Flask-Login>=0.6
Flask-WTF>=1.1.1
Flask-Session>=0.5.0
Flask-Limiter>=2.9
Flask-Talisman>=1.0.0
Flask-Bcrypt>=1.0.1
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
PyMySQL>=1.0
bleach>=6.0.0
